move files to my liking
