+++ app/models/user.server.ts
@@ -1,7 +1,7 @@
 import type { Password, User } from "@prisma/client";
 import bcrypt from "bcryptjs";
 
-import { prisma } from "~/db.server";
+import { prisma } from "~/services/db.server";
 
 export type { User } from "@prisma/client";
