+++ app/routes/logout.tsx
@@ -1,7 +1,7 @@
 import type { ActionFunctionArgs } from "@remix-run/node";
 import { redirect } from "@remix-run/node";
 
-import { logout } from "~/session.server";
+import { logout } from "~/services/session.server";
 
 export const action = async ({ request }: ActionFunctionArgs) =>
   logout(request);
