+++ app/services/singleton.server.ts
