+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "npm-run-all --sequential build:*",
     "build:remix": "remix build",
-    "build:server": "esbuild --platform=node --format=esm ./server.ts --outdir=build --bundle --external:fsevents",
+    "build:server": "esbuild --platform=node --format=esm server/server.ts --outdir=build --bundle --external:fsevents",
     "dev": "npm-run-all --parallel dev:*",
     "dev:server": "cross-env NODE_ENV=development npm run build:server -- --watch",
     "dev:remix": "remix dev --manual -c \"node --import ./mocks/index.js --watch-path ./build/server.js --watch ./build/server.js\"",
